refactor(backend): Unshare third-party resource from AppService
